FinalBackend-main/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-pymongo==2.3.0
flask-socketio==5.3.6
pymongo[srv]==3.12
werkzeug==3.0.3
pyjwt==2.8.0
requests==2.32.3
pytz==2024.1
joblib
numpy
scikit-learn
python-dotenv==1.0.1
gunicorn==22.0.0
dnspython==2.6.1
simple-websocket
